feat(site): support local preview
marked.js is used to render the corresponding raw markdown file
which is retrieved if the requested HTML file does not exists.

nginx in docker is also introduced to run the local server.

To make everything right, `/sitemap.xml` and `/assets/css/style.css`
should be downloaded to the local path.